refactor(handler): move to helpers module
